+++ Runtime/Scripts/TezosAPI/ITezosAPI.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Text.Json;
+using Beacon.Sdk.Beacon.Sign;
 using BeaconSDK;
 using TezosAPI.Models;
 using TezosAPI.Models.Tokens;
@@ -69,15 +70,14 @@ namespace TezosAPI
         /// <summary>
         /// Sends a request to the sign a payload string
         /// </summary>
-        /// <param name="signingType">enumerates the signing type (0 = MICHELINE, 1 = OPERATION, 2 = RAW)</param>
         /// <param name="payload">payload string that is going to be signed</param>
-        public void RequestSignPayload(int signingType, string payload);
+        public void RequestSignPayload(SignPayloadType signingType, string payload);
 
         /// <summary>
         /// Verify a signed payload to check if it is valid
         /// </summary>
         /// <param name="payload">payload string that is going to be signed</param>
-        public bool VerifySignedPayload(string payload);
+        public bool VerifySignedPayload(SignPayloadType signingType, string payload);
 
         /// <summary>
         /// Exposes a Monobehaviour class that exposes wallet events
@@ -86,7 +86,7 @@ namespace TezosAPI
 
         // Gets all tokens currently owned by a given address.
         public IEnumerator GetTokensForOwner(
-            Action<IEnumerable<TokenBalance>> cb,
+            Action<IEnumerable<TokenBalance>> callback,
             string owner,
             bool withMetadata,
             long maxItems,
@@ -94,7 +94,7 @@ namespace TezosAPI
 
         // Get the owner(s) for a token.
         public IEnumerator GetOwnersForToken(
-            Action<IEnumerable<TokenBalance>> cb,
+            Action<IEnumerable<TokenBalance>> callback,
             string contractAddress,
             uint tokenId,
             long maxItems,
@@ -102,33 +102,41 @@ namespace TezosAPI
 
         // Gets all owners for a given token contract.
         public IEnumerator GetOwnersForContract(
-            Action<IEnumerable<TokenBalance>> cb,
+            Action<IEnumerable<TokenBalance>> callback,
             string contractAddress,
             long maxItems,
             OwnersForContractOrder orderBy);
 
         // Checks whether a wallet holds a token in a given contract.
         public IEnumerator IsHolderOfContract(
-            Action<bool> cb,
+            Action<bool> callback,
             string wallet,
             string contractAddress);
         
         // Checks whether a wallet holds a particular token.
         public IEnumerator IsHolderOfToken(
-            Action<bool> cb,
+            Action<bool> callback,
             string wallet,
             string contractAddress,
-            string tokenId);
+            uint tokenId);
 
         // Gets the metadata associated with a given token.
         public IEnumerator GetTokenMetadata(
-            Action<JsonElement> cb,
+            Action<JsonElement> callback,
             string contractAddress,
             uint tokenId);
 
         // Queries token high-level collection/contract level information.
         public IEnumerator GetContractMetadata(
-            Action<JsonElement> cb,
+            Action<JsonElement> callback,
             string contractAddress);
+        
+        // Gets all tokens for a given token contract.
+        public IEnumerator GetTokensForContract(
+            Action<IEnumerable<Token>> callback,
+            string contractAddress,
+            bool withMetadata,
+            long maxItems,
+            TokensForContractOrder orderBy);
     }
 }