+++ CHANGELOG.md
@@ -5,13 +5,27 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased] - 2023-04-19
+## [Unreleased]
 ### Added
-- API methods: `GetOwnersForToken`, `GetOwnersForContract`, `IsHolderOfContract`, `GetTokenMetadata`, `GetContractMetadata`
-- Updated Dll binaries
+- API methods: `GetOwnersForToken`, `GetOwnersForContract`, `IsHolderOfContract`, `GetTokenMetadata`, `GetContractMetadata`, `GetTokensForContract`
+
+
+## [1.3.1] - 2023-04-27
+### Changed
+- Beacon SDK version to [1.0.23](https://github.com/baking-bad/beacon-dotnet-sdk/releases/tag/v1.0.23)
+
+
+## [1.3.0] - 2023-04-25
+### Added
+- Custom unity readme
 
 ### Fixed
+- Bug with unnecessary `RequestPermission()` call while clicking on `Deeplink connect` button on mobiles
+
+### Changed
 - Package structure
+- SignPayload method: now users can pass payload as plaintext without hex-encoding
+- Readme images
 
 
 ## [1.2.1] - 2023-04-17
@@ -53,7 +67,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added auto releases with GH actions
 
 
-[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/1.2.1...HEAD
+[unreleased]: https://github.com/trilitech/tezos-unity-sdk/compare/1.3.1...HEAD
+[1.3.1]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.3.1
+[1.3.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.3.0
 [1.2.1]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.2.1
 [1.2.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.2.0
 [1.1.0]: https://github.com/trilitech/tezos-unity-sdk/releases/tag/1.1.0